# docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /v2/members/{id}/relationships:
    get:
      operationId: RelationshipsController_getRelationships
      summary: Get all relationships for a member
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberRelationshipResponseDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/members/{id}/relationships/parents:
    get:
      operationId: RelationshipsController_getParents
      summary: Get parents of a member
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberRelationshipResponseDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/members/{id}/relationships/children:
    get:
      operationId: RelationshipsController_getChildren
      summary: Get children of a member
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberRelationshipResponseDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/members/{id}/relationships/spouses:
    get:
      operationId: RelationshipsController_getSpouses
      summary: Get spouses of a member
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberRelationshipResponseDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/members/{id}/relationships/ancestors:
    get:
      operationId: RelationshipsController_getAncestors
      summary: Get all ancestors (recursive)
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LineageMemberDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/members/{id}/relationships/descendants:
    get:
      operationId: RelationshipsController_getDescendants
      summary: Get all descendants (recursive)
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LineageMemberDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/relationships/search:
    get:
      operationId: RelationshipsController_searchRelationships
      summary: Search relationships by type, memberId, role
      parameters:
        - 
          name: memberId
          required: false
          in: query
          schema:
            type: string
        - 
          name: type
          required: false
          in: query
          schema:
            enum:
              - BIOLOGICAL
              - ADOPTED
              - SPOUSE
            type: string
        - 
          name: role
          required: false
          in: query
          schema:
            enum:
              - parent
              - child
              - spouse
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberRelationshipResponseDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/members/{memberId}/relationships:
    post:
      operationId: RelationshipsController_addRelationship
      summary: Add relationship between two members
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRelationshipDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberRelationshipResponseDto"
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/relationships/{id}:
    delete:
      operationId: RelationshipsController_deleteRelationship
      summary: Delete a relationship
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Relationships
      security:
        - 
          bearer: []
  /v2/tree/chart:
    get:
      operationId: TreeController_getChart
      summary: Get full family tree chart (Redis cached, 1h TTL)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTreeChartResponseDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
  /v2/tree/chart/{memberId}:
    get:
      operationId: TreeController_getSubTreeChart
      summary: Get 4-generation subtree from member
      parameters:
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTreeChartResponseDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
  /v2/tree/regenerate:
    post:
      operationId: TreeController_regenerate
      summary: Force regenerate tree chart + invalidate Redis cache
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTreeChartResponseDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
  /v2/tree/stats:
    get:
      operationId: TreeController_getStats
      summary: Get tree statistics + cache status
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponseDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
  /v2/tree/home:
    get:
      operationId: TreeController_getHomeTrees
      summary: Get trees with show=true (for homepage)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreeRecordDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
  /v2/tree:
    get:
      operationId: TreeController_getAllTrees
      summary: List all tree records
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreeRecordDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
    post:
      operationId: TreeController_createTree
      summary: Create new tree record (branch)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreeDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeRecordDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
  /v2/tree/{id}:
    get:
      operationId: TreeController_getTreeById
      summary: Get tree by ID
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeRecordDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
    put:
      operationId: TreeController_updateTree
      summary: Update tree record
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreeDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeRecordDto"
      tags:
        - Tree
      security:
        - 
          bearer: []
    delete:
      operationId: TreeController_deleteTree
      summary: Delete tree record
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Tree
      security:
        - 
          bearer: []
...
info:
  title: Family Tree API v2
  description: "## Vietnamese Family Tree Management API\n\n**Base URL:** `http://localhost:3002/v2`\n\n**Authentication:** Bearer JWT token (same as backend v1)\n\n### Modules\n- **Members** — CRUD thành viên + profile + avatar (async upload)\n- **Relationships** — Quan hệ mới (BIOLOGICAL/ADOPTED/SPOUSE) + tìm tổ tiên/con cháu\n- **Tree** — Cây gia phả full (Redis cache 1h) + subtree 4 thế hệ\n- **Anniversaries** — Ngày giỗ (filter by member, month, upcoming)\n- **Events** — Sự kiện dòng họ + notification queue\n- **Media** — Thư viện media: upload mọi loại (ảnh nén lossless bằng sharp; video/audio/tài liệu upload thẳng) → Vercel Blob; phân trang/lọc/tìm kiếm + thống kê + album\n- **Graves** — Mộ phần với tọa độ GPS + tìm kiếm gần nhất\n\n### Queue Jobs (QStash + Redis)\n| Queue | Trigger | Action |\n|-------|---------|--------|\n| avatar-upload | Create/Update member với file | Upload → Vercel Blob → cập nhật avatar_url |\n| media-process | Upload media | Ảnh: sharp nén lossless; khác: upload thẳng → Vercel Blob (off-request) |\n| report-generate | Create/Delete member | Tính stats → lưu Redis |\n| notification | New member/relationship/event | Log (Phase 1) |\n"
  version: 2.0.0
  contact:
    name: Family Tree Team
    url: ""
    email: ""
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
      description: Dán CHỈ token thô (lấy từ POST /v2/auth/login), KHÔNG kèm tiền tố "Bearer ".
  schemas:
...
    MemberTreeBriefDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: object
          nullable: true
          example: Chi nhánh Phú Thọ
      required:
        - id
...
    ProfileResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        fullName:
          type: string
          example: Nguyễn Văn A
        generation:
          type: number
          nullable: true
          example: 5
          description: Thế hệ NHẬP TAY — ưu tiên hơn giá trị suy ra. Xem `member.generation` cho giá trị hiệu lực.
        biography:
          type: object
          nullable: true
        occupation:
          type: object
          nullable: true
          example: Kỹ sư phần mềm
        address:
          type: object
          nullable: true
          example: Hà Nội
        notes:
          type: object
          nullable: true
        phone:
          type: object
          nullable: true
          example: 0988 123 456
        contactEmail:
          type: object
          nullable: true
          example: lienhe@gmail.com
        familyPosition:
          type: object
          nullable: true
          example: Con trưởng
          description: Vị trí trong gia đình
        roleTags:
          default: []
          description: Vai trò trong dòng họ
          type: array
          items:
            type: string
        committeeRole:
          type: object
          nullable: true
          description: Vai trò trong ban/hội đồng (clanRole)
        isCommittee:
          type: boolean
          default: false
        isNotable:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - member_id
        - fullName
        - roleTags
        - isCommittee
        - isNotable
        - created_at
        - updated_at
    MemberResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        private_id:
          type: object
          nullable: true
        name:
          type: string
          example: Nguyễn Văn A
        normalized_name:
          type: object
          nullable: true
          description: Tên đã bỏ dấu để tìm kiếm
        avatar_url:
          type: object
          nullable: true
          format: uri
        gender:
          type: string
          nullable: true
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: object
          nullable: true
          example: 1990-01-01
        deathDate:
          type: object
          nullable: true
          example: 2020-12-31
        tree_id:
          type: object
          nullable: true
          format: uuid
          description: Chi nhánh — FK tới Tree
        generation:
          type: number
          nullable: true
          example: 5
          description: "Thế hệ (giá trị HIỆU LỰC, đã tính sẵn) = `profile.generation` nếu nhập tay, ngược lại suy ra từ quan hệ cha-con/vợ-chồng. Cập nhật bởi job nền `generation-recompute`. Muốn biết là nhập tay hay suy ra: `profile.generation != null`."
        generation_updated_at:
          type: string
          nullable: true
          format: date-time
          description: Lần cuối job nền ghi lại `generation`
        created_at:
          type: string
          format: date-time
        profile:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/ProfileResponseDto"
        tree:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/MemberTreeBriefDto"
      required:
        - id
        - name
        - created_at
...
    MemberRelationshipResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - BIOLOGICAL
            - ADOPTED
            - SPOUSE
        note:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        parent:
          $ref: "#/components/schemas/MemberResponseDto"
        child:
          $ref: "#/components/schemas/MemberResponseDto"
      required:
        - id
        - parent_id
        - child_id
        - type
        - created_at
    LineageMemberDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gender:
          type: string
          nullable: true
          enum:
            - M
            - F
            - O
            - U
        avatar_url:
          type: object
          nullable: true
          format: uri
        depth:
          type: number
          example: 1
          description: Số thế hệ cách member gốc
      required:
        - id
        - name
        - depth
    CreateRelationshipDto:
      type: object
      properties:
        parentId:
          type: string
          example: uuid-parent
        childId:
          type: string
          example: uuid-child
        type:
          type: string
          enum:
            - BIOLOGICAL
            - ADOPTED
            - SPOUSE
        note:
          type: string
          example: Con nuôi từ năm 1990
      required:
        - parentId
        - childId
        - type
    FamilyChartRelsDto:
      type: object
      properties:
        spouses:
          type: array
          items:
            type: string
            format: uuid
        father:
          type: string
          format: uuid
        mother:
          type: string
          format: uuid
        children:
          type: array
          items:
            type: string
            format: uuid
    FamilyChartDataDto:
      type: object
      properties:
        gender:
          type: string
          enum:
            - M
            - F
            - O
            - U
        fn:
          type: string
          description: First name
        ln:
          type: string
          description: Last name / surname
        label:
          type: string
        birthday:
          type: string
        avatar:
          type: string
          format: uri
        generation:
          type: number
        desc:
          type: string
      required:
        - gender
        - fn
        - ln
        - label
    FamilyChartNodeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rels:
          $ref: "#/components/schemas/FamilyChartRelsDto"
        data:
          $ref: "#/components/schemas/FamilyChartDataDto"
      required:
        - id
        - rels
        - data
    FamilyTreeChartResponseDto:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/FamilyChartNodeDto"
        generatedAt:
          type: string
          format: date-time
      required:
        - nodes
        - generatedAt
    StatsResponseDto:
      type: object
      properties:
        totalMembers:
          type: number
          example: 120
        generations:
          type: number
          example: 5
        totalGenerations:
          type: number
          example: 5
          description: Alias của generations (backward-compat)
        deceased:
          type: number
          example: 30
          description: Số thành viên đã mất
        born20th21st:
          type: number
          example: 80
          description: Sinh trong khoảng 1901–2100
        lastUpdate:
          type: object
          nullable: true
          example: 2026-07-31
        generatedAt:
          type: string
          format: date-time
        cacheStatus:
          type: string
          enum:
            - hit
            - miss
          description: Trạng thái Redis cache
      required:
        - totalMembers
        - generations
        - totalGenerations
        - deceased
        - born20th21st
        - generatedAt
        - cacheStatus
    TreeRecordDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        title:
          type: object
          nullable: true
        description:
          type: object
          nullable: true
        image:
          type: object
          nullable: true
          format: uri
        show:
          type: boolean
          default: false
          description: Hiện trên trang chủ
        created_at:
          type: string
          format: date-time
      required:
        - id
        - owner_id
        - show
        - created_at
    CreateTreeDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        owner_id:
          type: string
          format: uuid
        show:
          type: boolean
          default: false
      required:
        - owner_id
    UpdateTreeDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        show:
          type: boolean
...