# src/relationships/relationships.controller.ts
import {
  Controller, Get, Post, Delete, Param, Body, Query, UseGuards, HttpCode, HttpStatus,
} from '@nestjs/common';
import {
  ApiTags, ApiBearerAuth, ApiOperation, ApiQuery,
  ApiOkResponse, ApiCreatedResponse, ApiNoContentResponse,
} from '@nestjs/swagger';
import { JwtAuthGuard } from '../auth/jwt.guard';
import { Roles } from '../auth/roles.decorator';
import { RolesGuard } from '../auth/roles.guard';
import { RelationshipsService } from './relationships.service';
import { CreateRelationshipDto, SearchRelationshipDto } from './dto/create-relationship.dto';
import {
  MemberRelationshipResponseDto,
  LineageMemberDto,
} from './dto/relationship-response.dto';

@ApiTags('Relationships')
@ApiBearerAuth()
@UseGuards(JwtAuthGuard, RolesGuard)
@Controller()
export class RelationshipsController {
  constructor(private readonly relationshipsService: RelationshipsService) {}

  @Get('members/:id/relationships')
  @ApiOperation({ summary: 'Get all relationships for a member' })
  @ApiOkResponse({ type: [MemberRelationshipResponseDto] })
  getRelationships(@Param('id') id: string) {
    return this.relationshipsService.getRelationships(id);
  }

  @Get('members/:id/relationships/parents')
  @ApiOperation({ summary: 'Get parents of a member' })
  @ApiOkResponse({ type: [MemberRelationshipResponseDto] })
  getParents(@Param('id') id: string) {
    return this.relationshipsService.getParents(id);
  }

  @Get('members/:id/relationships/children')
  @ApiOperation({ summary: 'Get children of a member' })
  @ApiOkResponse({ type: [MemberRelationshipResponseDto] })
  getChildren(@Param('id') id: string) {
    return this.relationshipsService.getChildren(id);
  }

  @Get('members/:id/relationships/spouses')
  @ApiOperation({ summary: 'Get spouses of a member' })
  @ApiOkResponse({ type: [MemberRelationshipResponseDto] })
  getSpouses(@Param('id') id: string) {
    return this.relationshipsService.getSpouses(id);
  }

  @Get('members/:id/relationships/ancestors')
  @ApiOperation({ summary: 'Get all ancestors (recursive)' })
  @ApiOkResponse({ type: [LineageMemberDto] })
  getAncestors(@Param('id') id: string) {
    return this.relationshipsService.getAncestors(id);
  }

  @Get('members/:id/relationships/descendants')
  @ApiOperation({ summary: 'Get all descendants (recursive)' })
  @ApiOkResponse({ type: [LineageMemberDto] })
  getDescendants(@Param('id') id: string) {
    return this.relationshipsService.getDescendants(id);
  }

  @Get('relationships/search')
  @ApiOperation({ summary: 'Search relationships by type, memberId, role' })
  @ApiOkResponse({ type: [MemberRelationshipResponseDto] })
  searchRelationships(@Query() query: SearchRelationshipDto) {
    return this.relationshipsService.searchRelationships(query);
  }

  @Post('members/:memberId/relationships')
  @Roles('editor')
  @ApiOperation({ summary: 'Add relationship between two members' })
  @ApiCreatedResponse({ type: MemberRelationshipResponseDto })
  addRelationship(@Body() dto: CreateRelationshipDto) {
    return this.relationshipsService.addRelationship(dto);
  }

  @Delete('relationships/:id')
  @Roles('admin')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ summary: 'Delete a relationship' })
  @ApiNoContentResponse({ description: 'Deleted' })
  deleteRelationship(@Param('id') id: string) {
    return this.relationshipsService.deleteRelationship(id);
  }
}